5/Services/Auth/requirements.txt
Flask
pyyaml
pyjwt
